fix: use equal-sized distinct dtypes in test_set_buffer_overlap_distinct
The test used float32 (16384B) and float16 (8192B) buffers in a distinct
reuse group, violating the constraint that units_between_buffer_groups
must be a multiple of each buffer's size. Changed to float16 + bfloat16
(both 8192B) which satisfies the size constraint while keeping different
element types to prevent CSE from merging the allocations.

+++ extensions/utlx/test/test_storage_alias.py
@@ -275,14 +275,14 @@ def test_set_buffer_overlap_distinct_compile_only():
         spec = tlx.storage_alias_spec(storage=tlx.storage_kind.smem)
         a = tlx.local_alloc(
             (BLOCK, BLOCK),
-            tl.float32,
+            tl.float16,
             tl.constexpr(2),
             tlx.storage_kind.smem,
             reuse=spec,
         )
         b = tlx.local_alloc(
             (BLOCK, BLOCK),
-            tl.float16,
+            tl.bfloat16,
             tl.constexpr(2),
             tlx.storage_kind.smem,
             reuse=spec,
